requirements.txt
ffmpeg-python == 0.2.0
nltk == 3.9.1
openai-whisper == 20250625
pydub == 0.25.1
python-docx == 1.2.0
streamlit == 1.49.0
torch == 2.8.0
datasets == 4.0.0
transformers == 4.55.4
torchvision == 0.23.0
numpy<2.3.0